have local player now reporting position status to the server

+++ shared/src/main/java/com/cormicopiastudios/theloftshared/SharedObjects/PlayerObject.java
@@ -1,7 +1,7 @@
 package com.cormicopiastudios.theloftshared.SharedObjects;
 
 public class PlayerObject {
-    public long tid;
+    public int tid;
     public String name;
     public PlayerPos pos;
 }
