+++ shared/src/main/java/com/cormicopiastudios/theloftshared/SharedObjects/PlayerPos.java
@@ -1,7 +1,7 @@
 package com.cormicopiastudios.theloftshared.SharedObjects;
 
 public class PlayerPos {
-    public long tid;
+    public int tid;
     public float x;
     public float y;
 }
